+++ app/api/rpcApi.js
@@ -136,8 +136,13 @@ function getBlockStatsByHeight(height) {
 	}
 }
 
-function getUtxoSetSummary() {
-	return getRpcData("gettxoutsetinfo");
+function getUtxoSetSummary(useCoinStatsIndexIfAvailable=true) {
+	if (useCoinStatsIndexIfAvailable && global.getindexinfo && global.getindexinfo.coinstatsindex) {
+		return getRpcDataWithParams({method:"gettxoutsetinfo", parameters:["muhash"]});
+
+	} else {
+		return getRpcData("gettxoutsetinfo");
+	}
 }
 
 function getRawMempool() {